# [SYCL] Optimize mul_mat for Q4_0 on Intel GPU