… docs/backend/SYCL.md
… examples/sycl/run-llama2.sh
… ggml/src/ggml-sycl/CMakeLists.txt
… ggml/src/ggml-sycl/getrows.cpp
… ggml/src/ggml-sycl/sycl_hw.cpp
… ggml/src/ggml-sycl/sycl_hw.hpp