Optimize MUL_MAT Q4_0 on Intel GPU.
- Change the number of threads of kernel function.
- Reorder the Q4 block to separate quantized weights and dequantize scaler.
   execute to reorder once during compute graph.
- Detect the Intel device type for optimize and save the result in ggml_backend_sycl_context.
- Show the result of detect hardware support optimize feature during startup.

```
  SYCL Optimization Feature:
  |ID|        Device Type|Reorder|
  |--|-------------------|-------|
  | 0| [level_zero:gpu:0]|      Y|
```
- Save the optimized status in tensor's extra.
- The optimize is only for Q4_0. The framework is easy to extend to QX_Y in the feature.
- Add environment variable GGML_SYCL_DISABLE_OPT to debug.
  It will be shown in startup:
  Running with Environment Variables:
```
  GGML_SYCL_DEBUG: 0
  GGML_SYCL_DISABLE_OPT: 0
```

It's passed in local CI.

Here is the performance increasing on Intel GPUs (dGPU and iGPU since MTL):
(Test with llama-2-7b.Q4_0.gguf)
|GPU|Base tokens/s|Increased tokens/s|
|-|-|-|
|PVC 1550|39|73|
|Flex 170|39|50|
|Arc770|42|55|
|MTL|13|16|
|ARL-H|14|17|

For iGPU which is older than MTL, the optimize doesn't increase performance. Skip to support them. Need more study.

For none Intel GPU, the GPU optimize feature detect doesn't support none Intel GPUs in code.
It's unknown the code is good on none Intel GPUs. It could be verified.

I hope this optimize is the seed of optimization on Intel GPU.
This solution is not the better solution on Intel GPU.
There is still huge potential of Intel GPU. Need more study work in the feature.

Thank @luoyu-intel and @airMeng for the solution contribution.