@qnixsynapse 2025-02-24
@Alcpz I am worried that this PR will break RPC(since it adds tensor->extra in the init tensor function again). I added a review but it was never attended for some reason.