@qnixsynapse 2025-02-24
I do not have multiple PCs with Intel GPUs unfortunately. But it was previous disabled because of the mentioned reason. Please see https://github.com/ggml-org/llama.cpp/discussions/5277#discussioncomment-11625312 .