@Alcpz 2025-02-24
@qnixsynapse I've confirmed that the RPC server is broken by this build. Thank you for your pointer:

```
❯ bin/rpc-server -p 50052
create_backend: using SYCL backend
Running with Environment Variables:
  GGML_SYCL_DEBUG: 0
  GGML_SYCL_DISABLE_OPT: 0
Build with Macros:
  GGML_SYCL_FORCE_MMQ: no
  GGML_SYCL_F16: yes
Found 1 SYCL devices:
|  |                   |                                       |       |Max    |        |Max  |Global |                     |
|  |                   |                                       |       |compute|Max work|sub  |mem    |                     |
|ID|        Device Type|                                   Name|Version|units  |group   |group|size   |       Driver version|
|--|-------------------|---------------------------------------|-------|-------|--------|-----|-------|---------------------|
| 0| [level_zero:gpu:0]|         Intel Data Center GPU Max 1100|  12.60|    448|    1024|   32| 51539M|         1.6.32224+14|
SYCL Optimization Feature:
|ID|        Device Type|Reorder|
|--|-------------------|-------|
| 0| [level_zero:gpu:0]|      Y|
get_memory_info: [warning] ext_intel_free_memory is not supported (export/set ZES_ENABLE_SYSMAN=1 to support), use total memory as free memory
Starting RPC server on 127.0.0.1:50052, backend memory: 49152 MB
Accepted client connection, free_mem=51539607552, total_mem=51539607552
Client connection closed
Accepted client connection, free_mem=51539607552, total_mem=51539607552
Client connection closed
Accepted client connection, free_mem=51539607552, total_mem=51539607552
Client connection closed
Accepted client connection, free_mem=51539607552, total_mem=51539607552
Client connection closed
Accepted client connection, free_mem=51539607552, total_mem=51539607552
./sources/llama.cpp-master/ggml/src/ggml-sycl/ggml-sycl.cpp:4120: GGML_ASSERT(extra) failed
[New LWP 2471816]
Registering SYCL extensions for gdb
[Thread debugging using libthread_db enabled]
Using host libthread_db library "/lib/x86_64-linux-gnu/libthread_db.so.1".
0x00007faece36842f in __GI___wait4 (pid=2472291, stat_loc=0x7fff0df43b5c, options=0, usage=0x0) at ../sysdeps/unix/sysv/linux/wait4.c:30
30      ../sysdeps/unix/sysv/linux/wait4.c: No such file or directory.
#0  0x00007faece36842f in __GI___wait4 (pid=2472291, stat_loc=0x7fff0df43b5c, options=0, usage=0x0) at ../sysdeps/unix/sysv/linux/wait4.c:30
30      in ../sysdeps/unix/sysv/linux/wait4.c
#1  0x00007faece7f8ae6 in ggml_print_backtrace () from ./sources/llama.cpp-master/build/bin/libggml-base.so
#2  0x00007faece7f8a80 in ggml_abort () from ./sources/llama.cpp-master/build/bin/libggml-base.so
#3  0x00007faece90a1ce in optimize_graph_once(ggml_cgraph*, ggml_backend_sycl_context*) () from ./sources/llama.cpp-master/build/bin/libggml-sycl.so
#4  0x00007faece92f509 in ggml_backend_sycl_graph_compute(ggml_backend*, ggml_cgraph*) () from sources/llama.cpp-master/build/bin/libggml-sycl.so
#5  0x00007faece80e437 in ggml_backend_graph_compute () from ./sources/llama.cpp-master/build/bin/libggml-base.so
#6  0x00007faecebfbfda in rpc_server::graph_compute(std::vector<unsigned char, std::allocator<unsigned char> > const&, rpc_msg_graph_compute_rsp&) () from ./sources/llama.cpp-master/build/bin/libggml-rpc.so
#7  0x00007faecebfcaf5 in ggml_backend_rpc_start_server () from ./sources/llama.cpp-master/build/bin/libggml-rpc.so
#8  0x0000000000404f42 in main ()
[Inferior 1 (process 2471814) detached]
[1]    2471814 IOT instruction  bin/rpc-server -p 50052
```

As you suspected, due to the extra field in the tensors. If I set `GGML_SYCL_DISABLE_OPT=1` I've confirmed that it works as expected, since we don't use the extra tensors that way. Despite the patch only breaking for models that use `q4_0` I think we should revert this a find a way to have these optimizations enabled without the use of the extra tensors.