@qnixsynapse 2025-02-25
@Alcpz Thank you so much for taking the initiative to test it.

> As you suspected, due to the extra field in the tensors. If I set GGML_SYCL_DISABLE_OPT=1 I've confirmed that it works as expected, since we don't use the extra tensors that way. Despite the patch only breaking for models that use q4_0 I think we should revert this a find a way to have these optimizations enabled without the use of the extra tensors.

Yes indeed.