@NeoZhangJianyu 2025-03-26
> @NeoZhangJianyu - why do you not use the SYCL joint_matrix extension here?

Joint_matrix is good for matrix * matrix.
In this case, it's matrix * vector.

Welcome you try with joint_matrix and contribute better code.